fix: shifted the rust-matrix-frame inside the pkg folder
