+++ numerix/.gitignore
@@ -1,3 +1,2 @@
 /target
 /Cargo.lock
-/test
