+++ numerix/env.example
@@ -7,4 +7,4 @@ APP_NAME="numerix"
 ETCD_SERVERS="127.0.0.1:2379"
 APP_LOG_LEVEL="INFO"
 CHANNEL_BUFFER_SIZE=10000
-LOG_SAPMLING_RATE=0.5
+LOG_SAMPLING_RATE=0.5
