--- numerix/numerix/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "numerix"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-prost = "0.11.9"
-prost-types = "0.11.9"
-tonic = "0.9"
-tokio = { version = "1.0", features = ["full"] }
-tokio-stream = { version = "0.1", features = ["net"] }
-hyper = { version = "0.14", features = ["full"] }
-tower = "0.4"
-tower-http = { version = "0.4", features = ["trace"] }
-config = "0.15.11"
-serde = "1.0.219"
-dotenv = "0.15.0"
-once_cell = "1.21.1"
-tracing = "0.1"
-tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
-futures = "0.3.31"
-serde_json = "1.0.140"
-cadence = "1.5.0"
-lazy_static = "1.4.0"
-zookeeper = "0.8"
-etcd-client = "0.14"
-rust-matrix-frame = {path = "../rust-matrix-frame/"}
-regex = "1.11.1"
-bytemuck = "1.14"
-ryu = "1.0"
-
-[build-dependencies]
-tonic-build = "0.10"
