+++ numerix/src/pkg/mod.rs
@@ -3,3 +3,4 @@ pub mod middleware;
 pub mod metrics;
 pub mod etcd;
 pub mod logger;
+pub mod rust_matrix_frame;
