+++ numerix/src/pkg/rust_matrix_frame/Cargo.toml
@@ -4,8 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-serde = { version = "1.0", features = ["derive"] }
-thiserror = "2.0.12"
+
 
 [lib]
 name = "rust_matrix_frame"
