+++ numerix/src/pkg/rust_matrix_frame/error.rs
