+++ numerix/src/pkg/rust_matrix_frame/mod.rs
@@ -1,0 +1,5 @@
+pub mod vector;
+pub mod ops;
+pub mod matrix;
+pub mod error;
+mod tests;
