+++ numerix/src/pkg/rust_matrix_frame/ops/mod.rs
@@ -1,6 +1,6 @@
-use crate::vector::Vector;
+use crate::pkg::rust_matrix_frame::vector::Vector;
 use serde::{Serialize, Deserialize};
-use crate::error::Mat2DError;
+use crate::pkg::rust_matrix_frame::error::Mat2DError;
 mod fp32_ops;
 mod fp64_ops;
